# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Balance Tracker API
  version: "1.0"
  description: >
    Household finance tracker API. This spec covers every route in
    `server/index.js` — auth, household/settings, transactions (+ trash),
    accounts, goals, positions, trips (+ expenses/sharing/guest access),
    backup/restore, widget, calendar feed, and admin (blacklist + metrics).
    It is a snapshot maintained by hand alongside the code (MOD-024), not
    generated. See ./API.md for pagination, idempotency, rate limiting,
    and the full error-code reference this spec references but doesn't
    repeat, plus the "Known inconsistencies" notes on error/response-shape
    drift across older routes.
servers:
  - url: /api
    description: Same-origin (production, via the Vercel proxy) or http://localhost:3001/api in development
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: httpOnly JWT session cookie, set by /auth/login or /auth/register.
    capabilityTokenQuery:
      type: apiKey
      in: query
      name: key
      description: Widget/calendar capability token — unrelated to household auth, grants only the narrow surface it was issued for.
    adminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Must match the server's ADMIN_SECRET environment variable. Unrelated to household auth.
  schemas:
    PortfolioSnapshot:
      type: object
      required: [weekKey, date, valore, investito, holdings]
      properties:
        weekKey: { type: string, format: date, description: The Sunday on or before the snapshot date (local calendar). }
        date: { type: string, format: date }
        valore: { type: number, description: Total market value (quantity × price) at the time. }
        investito: { type: number, description: Cost basis of the open holdings. }
        holdings:
          type: array
          items:
            type: object
            required: [ticker, quantita, prezzo]
            properties:
              ticker: { type: string }
              quantita: { type: number }
              prezzo: { type: number }
...
    Position:
      type: object
      description: >
        One buy or sell trade. Portfolio value/cost-basis is computed
        client- and server-side from the full set of a household's
        Positions per ticker using single average-cost accounting — see
        "Cost-basis accounting (MOD-018)" in API.md for the full method,
        partial/full-sale/oversell behavior, and known limitations
        (no fees, no cross-currency conversion despite the valuta field
        below). This is portfolio tracking, not tax accounting — no
        jurisdiction's tax-lot rules are modeled.
      properties:
        id: { type: string }
        ticker: { type: string, maxLength: 20 }
        nome: { type: string }
        quantita: { type: number }
        prezzoAcquisto: { type: number }
        prezzoAcquistoMinorUnits: { type: integer, description: "Integer-cents companion of prezzoAcquisto (MOD-016)." }
        dataAcquisto: { type: string, format: date }
        valuta:
          type: string
          default: EUR
          description: >
            Recorded but NOT used in cost-basis/value calculations today —
            every position is summed as if it were the household's base
            currency regardless of this field (MOD-018 known limitation).
        note: { type: string }
        tipo: { type: string, enum: [buy, sell], default: buy }
        contoId:
          type: string
          nullable: true
          description: >
            Household account this position's cash moved to/from. Setting
            it creates/updates a linked transaction (uscita for a buy,
            entrata for a sell, categoria "investimenti") so the account's
            balance reflects the purchase; clearing it deletes that
            transaction. See "Funding a position from an account" in API.md.
        linkedTransactionId:
          type: string
          readOnly: true
          description: Id of the transaction linked via contoId, if any. Server-managed — not settable directly.
      required: [id, ticker, quantita, prezzoAcquisto]
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 100 }
      description: See API.md#idempotency. Recommended for all create requests.
...
paths:
...
  /positions:
    get:
      summary: List portfolio positions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Position" } }
    post:
      summary: Record a buy/sell position
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Position" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Position" }
  /positions/{id}:
    put:
      summary: Partially update a position (e.g. rename its ticker)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Position" }
      responses:
        "200": { description: OK }
    delete:
      summary: Delete a position
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted. }
  /positions/prices:
    get:
      summary: Get manual price overrides
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  manualPrices: { type: object, additionalProperties: { type: number } }
    put:
      summary: Set manual price overrides
      description: Incremental upsert — only sets/updates the ticker(s) named in the request body; any other ticker's already-saved manual price is left untouched. Send only what changed, not the whole local snapshot.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manualPrices]
              properties:
                manualPrices:
                  type: object
                  additionalProperties: { type: number }
                  maxProperties: 200
      responses:
        "200": { description: "{ ok: true }." }
  /positions/prices/{ticker}:
    delete:
      summary: Clear one ticker's manual price override
      description: Removes only the named ticker's manual override; every other ticker's saved price is unaffected. Returns 200 even if the ticker had no manual price set.
      parameters:
        - name: ticker
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "{ ok: true }." }
        "400": { description: Malformed ticker. }
  /quotes:
    get:
      summary: Live stock/fund quote lookup
      description: Best-effort lookup against Yahoo Finance, cached server-side for 24h per ticker. A ticker that can't be resolved is omitted from the response rather than causing an error.
      parameters:
        - name: tickers
          in: query
          required: true
          schema: { type: string }
          description: Comma-separated ticker list (max 30), Yahoo syntax (e.g. `AAPL,BTC-USD,^GSPC,SWDA.MI`).
        - name: force
          in: query
          required: false
          schema: { type: string, enum: ["1", "true"] }
          description: Bypass the 24h cache and re-fetch every requested ticker.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  quotes:
                    type: object
                    additionalProperties: { type: number }
...
  /portfolio/snapshots:
    get:
      summary: List the weekly portfolio value snapshots
      description: Sorted by date ascending. Each snapshot is one real record of quantity × price per holding taken on the first price refresh of a Sunday-started week.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items: { $ref: "#/components/schemas/PortfolioSnapshot" }
    post:
      summary: Record this week's portfolio snapshot
      description: First write per (household, weekKey) wins — a later save for the same week is a no-op and returns 200 with created false. Requires at least the member role when a persona is logged in.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortfolioSnapshot" }
      responses:
        "201": { description: "Recorded. { created: true, weekKey }." }
        "200": { description: "A snapshot for that week already existed. { created: false, weekKey }." }
        "400": { description: Validation error (weekKey must be a Sunday, date inside that week, positive amounts, 1-200 holdings). }
...